requirements.txt
django
psycopg2-binary
django-guardian
django[argon2]
django-markdownx
django_multitenant
django-admin-json-editor
django-object-actions
opencensus-ext-azure
opencensus-ext-logging
opencensus-ext-requests
opencensus-ext-django
opencensus-ext-postgresql
gunicorn
uvicorn[standard]
django-storages[azure]
azure-core
azure-common
azure-storage-common
azure-storage-blob==12.8.0
